% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeIndegreeReceiver}
\alias{activeIndegreeReceiver}
\title{activeIndegreeReceiver}
\usage{
activeIndegreeReceiver(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeIndegreeReceiver}}.
}
