% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeOutdegreeSender}
\alias{activeOutdegreeSender}
\title{activeOutdegreeSender}
\usage{
activeOutdegreeSender(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeOutdegreeSender}}.
}
