% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeSharedPartners_isp}
\alias{activeSharedPartners_isp}
\title{activeSharedPartners_isp}
\usage{
activeSharedPartners_isp(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeSharedPartners_isp}}.
}
