% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeTotaldegreeDyad}
\alias{activeTotaldegreeDyad}
\title{activeTotaldegreeDyad}
\usage{
activeTotaldegreeDyad(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeTotaldegreeDyad}}.
}
