% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeTotaldegreeReceiver}
\alias{activeTotaldegreeReceiver}
\title{activeTotaldegreeReceiver}
\usage{
activeTotaldegreeReceiver(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeTotaldegreeReceiver}}.
}
