% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{actor_effects}
\alias{actor_effects}
\title{actor_effects}
\usage{
actor_effects(step = NULL)
}
\arguments{
\item{step}{outputs all statistics in the sender activity step (if `step =
sender`) or receiver choice step (if `step = receiver`).}
}
\description{
See \code{\link[remstats]{actor_effects}}.
}
