% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstimate.R
\name{bic_table}
\alias{bic_table}
\title{bic_table}
\usage{
bic_table(x, ...)
}
\arguments{
\item{x}{A \code{remstimate_mixrem_list} returned when \code{k} is a
vector in \code{remstimate(..., method = "MIXREM")}.}

\item{...}{Unused.}
}
\description{
See \code{\link[remstimate]{bic_table}}.
}
