% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist0.R
\docType{data}
\name{edgelist0}
\alias{edgelist0}
\title{Simulated relational event history}
\format{
A dataframe with 1000 rows and 3 variables:
\describe{
\item{time}{time of the event}
\item{actor1}{the first actor involved in the event}
\item{actor2}{the second actor involved in the event}
\item{setting}{the setting of an event: \code{X} or \code{Y}}
}
}
\source{
Simulated relational event sequence among 5 actors in a social network.
}
\usage{
data(edgelist0)
}
\description{
A simulated event sequence among 5 actors using the endogenous effects:
inertia, reciprocity, and itp.
}
\examples{
data(edgelist0)

}
\keyword{dataset}
