% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist0.R
\docType{data}
\name{edgelist0_actors}
\alias{edgelist0_actors}
\title{Simulated relational event history}
\format{
A dataframe with 1000 rows and 3 variables:
\describe{
\item{name}{label of the actor}
\item{time}{the time of measurement of the attribute}
\item{job}{the job of the actor}
}
}
\source{
Simulated attributes of 5 actors in a social network.
}
\usage{
data(edgelist0_actors)
}
\description{
Actor attributes of actors in event sequence \code{edgelist0}
}
\examples{
data(edgelist0_actors)

}
\keyword{dataset}
