% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist_duration.R
\docType{data}
\name{edgelist_duration}
\alias{edgelist_duration}
\title{Example relational event history with duration}
\format{
A data frame with columns:
\describe{
\item{time}{Event start time.}
\item{actor1}{Sender.}
\item{actor2}{Receiver.}
\item{end_time}{Event end time.}
}
}
\usage{
data(edgelist_duration)
}
\description{
A small directed relational event history where each event has a known
start and end time, suitable for duration modeling.
}
\examples{
data(edgelist_duration)
reh <- remify(edgelist_duration, duration = TRUE)

}
\keyword{datasets}
