% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{event}
\alias{event}
\title{event}
\usage{
event(variable, event_attr)
}
\arguments{
\item{variable}{string with the name of the column in the event_attr object for which the statistic has to be computed.}

\item{event_attr}{an object of class data.frame that contains the attribute}
}
\description{
See \code{\link[remstats]{event}}.
}
