% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstimate.R
\name{frailty_rem}
\alias{frailty_rem}
\title{frailty_rem}
\usage{
frailty_rem(reh, stats, ...)
}
\arguments{
\item{reh}{A \code{remify} or \code{remify_durem} object.}

\item{stats}{A \code{remstats} object (\code{tomstats}, \code{aomstats},
or \code{remstats_durem}).}

\item{...}{Additional arguments passed to \code{\link[remstimate]{remstimate}}.}
}
\description{
See \code{\link[remstimate]{frailty_rem}}.
}
