% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{indegreeSender}
\alias{indegreeSender}
\title{indegreeSender}
\usage{
indegreeSender(scaling = c("none", "prop", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{indegreeSender}}.
}
