% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{is.remstats_durem}
\alias{is.remstats_durem}
\title{is.remstats_durem}
\usage{
is.remstats_durem(x)
}
\arguments{
\item{x}{Any R object.}
}
\description{
See \code{\link[remstats]{is.remstats_durem}}.
}
