% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{psABBY}
\alias{psABBY}
\title{psABBY}
\usage{
psABBY(consider_type = TRUE)
}
\arguments{
\item{consider_type}{logical, indicates whether to consider the event type
in determining which dyads create a pshift (TRUE, default) or not (FALSE).}
}
\value{
List with all information required by `remstats::remstats()` to compute the statistic.
}
\description{
Specifies the statistic for a participation shift AB-BY.
}
\details{
The AB-BY participation shift refers to one of Gibson's (2003) dyadic participation shifts. The AB-BY pshift refers to a tendency for turn receiving (here, the next sender is the previous receiver and the next receiver is not in the current previous). For each timepoint t, the psABBY statistic is equal to one for the dyads that will create the participation shift if they would occur in the edgelist at time t and equal to zero for the dyads that will not create this participation shift. If consider_type is set to TRUE, the type of the AB event and the type of the BY events have to be equal. If it is set to FALSE, the participation shift is set to one for every BY event, regardless of the event type. If multiple events in the edgelist occur at the same time point, the order of these events determines whether the p-shift is observed. Note that the AB-BY pshift is only defined for directed events.
}
\examples{
reh_tie <- remify::remify(history, model = "tie")
effects <- ~ psABBY()
remstats(reh = reh_tie, tie_effects = effects)
}
\seealso{
\code{\link[remstats]{psABBA}}, \code{\link[remstats]{psABXA}}, \code{\link[remstats]{psABXB}},
\code{\link[remstats]{psABXY}} or \code{\link[remstats]{psABAY}} for other dyadic participation
shifts.
}
