% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{recencyReceiveSender}
\alias{recencyReceiveSender}
\title{recencyReceiveSender}
\usage{
recencyReceiveSender(consider_type = "ignore")
}
\description{
See \code{\link[remstats]{recencyReceiveSender}}.
}
