% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstimate.R
\name{rempenalty}
\alias{rempenalty}
\title{rempenalty}
\usage{
rempenalty(
  reh,
  stats,
  approach = c("frequentist", "Bayesian"),
  alpha = 1,
  prior = "horseshoe",
  nfolds = 10L,
  lambda_select = c("1se", "min"),
  ...
)
}
\description{
See \code{\link[remstimate]{rempenalty}}.
}
