% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remverse-package.R
\docType{package}
\name{remverse-package}
\alias{remverse}
\alias{remverse-package}
\title{remverse: A Collection of R Packages for Relational Event Modeling}
\description{
The remverse package loads an ensemble of R packages for working with
Relational Event Histories (REH): preprocessing event sequences
(\pkg{remify}), computing network statistics (\pkg{remstats}), and
estimating relational event models (\pkg{remstimate}).
}
\examples{
library(remverse)

# Load example data
data("randomREHsmall")

# Preprocess
reh <- remify(edgelist = randomREHsmall$edgelist,
              model = "tie",
              directed = TRUE)

# Compute statistics
stats <- remstats(reh,
                  tie_effects = ~ inertia(scaling="std") + reciprocity(scaling="std"),
                  start = 10)

# Fit model
fit <- remstimate(reh, stats)
summary(fit)

# Check diagnostics
diag_fit <- diagnostics(fit, reh, stats)
print(diag_fit)
plot(fit, reh, diag_fit)

}
\author{
\strong{Maintainer}: Joris Mulder \email{j.mulder3@tilburguniversity.edu}

Authors:
\itemize{
  \item Giuseppe Arena \email{g.arena@tilburguniversity.edu}
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu}
  \item Marlyne Meijerink-Bosman \email{m.l.meijerink@tilburguniversity.edu}
  \item Rumana Lakdawala \email{r.j.lakdawala@tilburguniversity.edu}
  \item Fabio Generoso Vieira \email{f.v.generosovieira@tilburguniversity.edu}
}

Other contributors:
\itemize{
  \item Mahdi Shafiee Kamalabad \email{m.shafiee@tilburguniversity.edu} [contributor]
  \item Diana Karimova \email{d.karimova@tilburguniversity.edu} [contributor]
}

}
