% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstimate.R
\name{waic}
\alias{waic}
\title{waic}
\usage{
waic(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object.}

\item{...}{further arguments to be passed to the 'waic' method.}
}
\value{
WAIC value of a 'remstimate' object.
}
\description{
A function that returns the WAIC (Watanabe-Akaike's Information Criterion) value in a 'remstimate' object.
}
\section{Methods (by class)}{

\itemize{
\item \code{waic(remstimate)}: WAIC (Watanabe-Akaike's Information Criterion) value of a 'remstimate' object

}
}

\examples{
# No examples available at the moment
}
