% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeDegreeDyad}
\alias{activeDegreeDyad}
\title{activeDegreeDyad}
\usage{
activeDegreeDyad(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeDegreeDyad}}.
}
