% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeDegreeMax}
\alias{activeDegreeMax}
\title{activeDegreeMax}
\usage{
activeDegreeMax(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeDegreeMax}}.
}
