% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeDegreeMin}
\alias{activeDegreeMin}
\title{activeDegreeMin}
\usage{
activeDegreeMin(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeDegreeMin}}.
}
