% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeReciprocalTie}
\alias{activeReciprocalTie}
\title{activeReciprocalTie}
\usage{
activeReciprocalTie(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeReciprocalTie}}.
}
