% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeSharedPartners}
\alias{activeSharedPartners}
\title{activeSharedPartners}
\usage{
activeSharedPartners(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeSharedPartners}}.
}
