% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeSharedPartners_itp}
\alias{activeSharedPartners_itp}
\title{activeSharedPartners_itp}
\usage{
activeSharedPartners_itp(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeSharedPartners_itp}}.
}
