% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeSharedPartners_osp}
\alias{activeSharedPartners_osp}
\title{activeSharedPartners_osp}
\usage{
activeSharedPartners_osp(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeSharedPartners_osp}}.
}
