% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeSharedPartners_otp}
\alias{activeSharedPartners_otp}
\title{activeSharedPartners_otp}
\usage{
activeSharedPartners_otp(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeSharedPartners_otp}}.
}
