% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{activeTotaldegreeSender}
\alias{activeTotaldegreeSender}
\title{activeTotaldegreeSender}
\usage{
activeTotaldegreeSender(scaling = c("none", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{activeTotaldegreeSender}}.
}
