% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstimate.R
\name{AICC}
\alias{AICC}
\title{AICC}
\usage{
AICC(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object.}

\item{...}{further arguments to be passed to the 'AICC' method.}
}
\description{
See \code{\link[remstimate]{AICC}}.
}
