% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{difference}
\alias{difference}
\title{difference}
\usage{
difference(
  variable,
  attr_actors = NULL,
  scaling = c("none", "std"),
  absolute = TRUE,
  attr_data
)
}
\description{
See \code{\link[remstats]{difference}}.
}
