% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{difference}
\alias{difference}
\title{difference}
\usage{
difference(
  variable,
  attr_actors = NULL,
  scaling = c("none", "std"),
  absolute = TRUE,
  attr_data
)
}
\arguments{
\item{variable}{string with the name of the column in the
\code{attr_actors} object for which the statistic has to be computed.}

\item{attr_actors}{optionally, an object of class
\code{\link[base]{data.frame}} that contains the attribute, see
'Details.'}

\item{scaling}{the method for scaling the statistic. Default is to not scale
the statistic. Alternatively, standardization of the
statistic per time point can be requested with "std".}

\item{absolute}{Logical value indicating whether the difference values
should be converted to the absolute difference (default is TRUE).}

\item{attr_data}{Deprecated argument. Please use 'attr_actors' instead.}
}
\description{
See \code{\link[remstats]{difference}}.
}
