% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\docType{data}
\name{randomREH3}
\alias{randomREH3}
\title{Generated Relational Event History with Duration, Type, and Weight}
\format{
\code{data(randomREH3)} loads a \code{data.frame} with 999 rows and
6 columns:
\describe{
\item{\code{time}}{the timestamp indicating when each event started;}
\item{\code{actor1}}{the actor that generated (initiated) the event;}
\item{\code{actor2}}{the actor that received the event;}
\item{\code{end}}{the time at which the event ended (\code{NA} = right-censored);}
\item{\code{setting}}{the setting (type) of the event, either \code{social} or \code{work};}
\item{\code{duration}}{the duration of each event (\code{end - time}), which can also serve as an event weight.}
}
}
\usage{
data(randomREH3)
}
\description{
A randomly generated sequence of 999 duration relational events among 5
actors. Starting an event depended on inertia, reciprocity, sender
out-degree, incoming shared partners, and sender age; (undirected) ending an
event depended on the dyad's total active degree and the age difference of
the two actors. Exponential memory decay was used with a half-life of 2000.
Actor covariates are in \code{\link{info3}}.
}
\seealso{
\code{\link{info3}} for actor covariates.
}
\keyword{datasets}
