% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{recencyReceiveReceiver}
\alias{recencyReceiveReceiver}
\title{recencyReceiveReceiver}
\usage{
recencyReceiveReceiver(consider_type = "ignore")
}
\description{
See \code{\link[remstats]{recencyReceiveReceiver}}.
}
