% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{recencySendReceiver}
\alias{recencySendReceiver}
\title{recencySendReceiver}
\usage{
recencySendReceiver(consider_type = "ignore")
}
\description{
See \code{\link[remstats]{recencySendReceiver}}.
}
