% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstimate.R
\name{remfrailty}
\alias{remfrailty}
\title{remfrailty}
\usage{
remfrailty(
  reh,
  stats,
  approach = c("frequentist", "Bayesian"),
  engine = "auto",
  ...
)
}
\description{
See \code{\link[remstimate]{remfrailty}}.
}
