% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstimate.R
\name{remtribute}
\alias{remtribute}
\title{remtribute}
\usage{
remtribute(
  reh,
  stats = NULL,
  effects = NULL,
  attribute = "type",
  attribute_type = c("nominal", "ordinal", "numeric"),
  attr_actors = NULL,
  memory = "full",
  memory_value = Inf,
  ...
)
}
\description{
See \code{\link[remstimate]{remtribute}}.
}
