% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstimate.R
\name{remwindow}
\alias{remwindow}
\title{remwindow}
\usage{
remwindow(
  reh,
  stats,
  n.windows = 5L,
  window.width = NULL,
  step.size.window = NULL,
  start.point = 1L,
  min.events = 50L,
  approach = c("frequentist", "Bayesian"),
  parallel = FALSE,
  ncores.window = 1L,
  ...
)
}
\description{
See \code{\link[remstimate]{remwindow}}.
}
