% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{same}
\alias{same}
\title{same}
\usage{
same(variable, attr_actors = NULL, attr_data)
}
\arguments{
\item{variable}{string with the name of the column in the
\code{attr_actors} object for which the statistic has to be computed.}

\item{attr_actors}{optionally, an object of class
\code{\link[base]{data.frame}} that contains the attribute, see
'Details.'}

\item{attr_data}{Deprecated argument. Please use 'attr_actors' instead.}
}
\description{
See \code{\link[remstats]{same}}.
}
