% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{select_stats}
\alias{select_stats}
\title{select_stats}
\usage{
select_stats(
  object,
  tie_effects = NULL,
  sender_effects = NULL,
  receiver_effects = NULL,
  start_effects = NULL,
  end_effects = NULL
)
}
\arguments{
\item{object}{a \code{remstats} object (of class \code{tomstats},
\code{aomstats}, or \code{remstats_durem})}

\item{tie_effects}{character vector of statistic names to keep (tomstats only)}

\item{sender_effects}{character vector of statistic names to keep (aomstats only)}

\item{receiver_effects}{character vector of statistic names to keep (aomstats only)}

\item{start_effects}{character vector of statistic names to keep (remstats_durem only)}

\item{end_effects}{character vector of statistic names to keep (remstats_durem only)}
}
\description{
See \code{\link[remstats]{select_stats}}.
}
