% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstats.R
\name{totaldegreeDyad}
\alias{totaldegreeDyad}
\title{totaldegreeDyad}
\usage{
totaldegreeDyad(scaling = c("none", "prop", "std"), consider_type = "ignore")
}
\description{
See \code{\link[remstats]{totaldegreeDyad}}.
}
