% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports-remstimate.R
\name{WAIC}
\alias{WAIC}
\title{WAIC}
\usage{
WAIC(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object.}

\item{...}{further arguments to be passed to the 'WAIC' method.}
}
\description{
See \code{\link[remstimate]{WAIC}}.
}
